requirements.txt
absl-py
aiohttp
aiosignal
appdirs==1.4.4
appnope
asttokens
astunparse==1.6.3
async-timeout
attrs
backcall
backports.functools-lru-cache
blinker==1.4
Bottleneck
brotlipy==0.7.0
cachetools
certifi
cffi
charset-normalizer
click
contourpy==1.0.6
cryptography
cycler==0.11.0
debugpy
decorator
executing
Flask
flatbuffers
fonttools==4.38.0
frozenlist
gast
google-auth
google-auth-oauthlib
google-pasta
grpcio
h5py
idna
importlib-metadata
ipykernel
ipython
itsdangerous
jedi 
Jinja2
jupyter_client
jupyter_core
keras
Keras-Preprocessing
kiwisolver==1.4.4
Markdown
MarkupSafe
matplotlib==3.5.3
matplotlib-inline
multidict
nest-asyncio
numexpr
numpy
oauthlib
opencv-python-headless
opt-einsum
packaging
pandas==1.3.5
parso
pexpect
pickleshare
Pillow==9.5.0
platformdirs
pooch 
prompt-toolkit
protobuf==3.20.3
psutil 
ptyprocess
pure-eval 
pyasn1 
pyasn1-modules==0.2.8
pycparser 
pydicom==2.3.1
Pygments
PyJWT 
pylibjpeg==1.4.0
pyOpenSSL 
pyparsing==3.0.9
PySocks
python-dateutil
python-gdcm==3.0.21
pytz 
pyzmq 
requests 
requests-oauthlib==1.3.0
rsa
scipy==1.7.3
six 
stack-data 
tensorboard
tensorboard-data-server 
tensorboard-plugin-wit
tensorflow 
tensorflow-estimator 
termcolor
tornado 
traitlets
typing_extensions 
urllib3 
wcwidth 
Werkzeug 
wrapt 
yarl 
zipp
gunicorn
firebase-admin
pylibjpeg-libjpeg
google-cloud-storage